refactor: scenariofunctions.js, general.js
