+++ panels/createpostmancollection/general.js
@@ -1,0 +1,8 @@
+export function isEmpty(string) {
+    var empty = false;
+    if (string === '' || string === undefined || string === null) {
+      empty = true;
+    }
+  
+    return empty;
+}
